Trouble shoot audio player

+++ db/migrate/20140220152828_create_tracks.rb
@@ -2,8 +2,8 @@ class CreateTracks < ActiveRecord::Migration
   def change
     create_table :tracks do |t|
       t.integer :user_id
-      t.integer :longitude
-      t.integer :latitude
+      t.float :longitude  # stored as ruby object BigDecimal in database - to access longitude call .to_s
+      t.float :latitude
       t.string :title
       t.text :description
       t.string :url
